Solaris 9 does not have ceilf(), so change the ListServerLink constructor to take a long, change ListServerReAddTime from float to double to better match the way it is used, and then good old ceil() is the correct function to use when calling the refactored constructor.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22790 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/ListServerConnection.h
@@ -32,7 +32,7 @@ public:
     // c'tor will fill list and local server information variables and
     // do an initial ADD
     ListServerLink(std::string listServerURL, std::string publicizedAddress,
-		   std::string publicizedTitle, std::string advertiseGroups, float dnsCache = 0);
+		   std::string publicizedTitle, std::string advertiseGroups, long dnsCache = 0);
     // c'tor with no arguments called when we don't want to use a list server.
     ListServerLink();
     // d'tor will REMOVE server and close connection
